sdk/requirements-test.txt
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.3.0
requests>=2.28.0
pytest>=7.0.0
pytest-cov>=4.0.0 